% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstate.R
\name{AIC.fmsm}
\alias{AIC.fmsm}
\title{Akaike's information criterion from a flexible parametric multistate model}
\usage{
\method{AIC}{fmsm}(object, ..., k = 2)
}
\arguments{
\item{object}{Object returned by \code{\link{fmsm}} representing a multistate model.}

\item{...}{Further arguments (currently unused).}

\item{k}{Penalty applied to number of parameters (defaults to the standard 2).}
}
\value{
The sum of the AICs of the transition-specific models.
}
\description{
Defined as the sum of the AICs of the transition-specific models.
}
