% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexsurvreg.R
\name{AICc}
\alias{AICc}
\alias{AICC}
\title{Second-order Akaike information criterion}
\usage{
AICc(object, ...)

AICC(object, ...)
}
\arguments{
\item{object}{Fitted model object.}

\item{...}{Other arguments (currently unused).}
}
\description{
Generic function for the second-order Akaike information criterion.
The only current implementation of this in \pkg{flexsurv} is
in \code{\link{AICc.flexsurvreg}}, see there for more details.
}
\details{
This can be spelt either as \code{AICC} or \code{AICc}.
}
