% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.hess_to_cov}
\alias{.hess_to_cov}
\title{helper function to safely convert a Hessian matrix to covariance matrix}
\usage{
.hess_to_cov(hessian, tol.solve = 0.000000001, tol.evalues = 0.00001, ...)
}
\arguments{
\item{hessian}{hessian matrix to convert to covariance matrix (must be evaluated at MLE)}

\item{tol.solve}{tolerance used for solve()}

\item{tol.evalues}{accepted tolerance for negative eigenvalues of the covariance matrix}

\item{...}{arguments passed to Matrix::nearPD}
}
\description{
helper function to safely convert a Hessian matrix to covariance matrix
}
\keyword{internal}
