% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.hessian}
\alias{.hessian}
\title{Numerical evaluation of the hessian of a function using numDeriv::hessian}
\usage{
.hessian(f, x, seconds.warning = 60, default.r = 6, min.r = 2, ...)
}
\arguments{
\item{f}{function to compute Hessian for}

\item{x}{location to evaluate Hessian at}

\item{seconds.warning}{time threshold in seconds to trigger message and
reduce the number of iterations for Richardson extrapolation of
numDeriv::hessian}

\item{default.r}{default number of iterations (high-precision recommendation
of numDeriv)}

\item{min.r}{minial number of iteration, must be at least 2,}

\item{...}{further arguments passed to method.args of numDeriv::hessian}
}
\description{
We perform a quick check about the expected runtime and adjust the
precision accordingly.
}
\keyword{internal}
