% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmixmsm.R
\name{fmixmsm}
\alias{fmixmsm}
\title{Constructor for a mixture multi-state model based on flexsurvmix}
\usage{
fmixmsm(...)
}
\arguments{
\item{...}{Named arguments.  Each argument should be a fitted model as
returned by \code{\link{flexsurvmix}}.  The name of each argument names
the starting state for that model.}
}
\value{
A list of \code{\link{flexsurvmix}} objects, with the following
  attribute(s):

  \code{pathways} A list of all potential pathways until absorption, for
  models without cycles.   For models with cycles this will have an element
  \code{has_cycle=TRUE}, plus the pathways discovered before the function
  found the cycle and gave up.
}
\description{
Constructor for a mixture multi-state model based on flexsurvmix
}
