% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.flexsurvmix.R
\name{get_basepars}
\alias{get_basepars}
\title{Evaluate baseline time-to-event distribution parameters given covariate values in a flexsurvmix model}
\usage{
get_basepars(x, newdata, event)
}
\arguments{
\item{x}{Fitted model object}

\item{newdata}{Data frame of alternative covariate values}

\item{event}{Event}
}
\value{
A list with one component per parameter, that gives values of that parameter for the different covariate values
}
\description{
Evaluate baseline time-to-event distribution parameters given covariate values in a flexsurvmix model
}
