% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexsurvreg.R
\name{logLik.flexsurvreg}
\alias{logLik.flexsurvreg}
\title{Log likelihood from a flexsurvreg model}
\usage{
\method{logLik}{flexsurvreg}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class
\code{\link{flexsurvreg}}, e.g. as returned by
\code{flexsurvreg} or \code{flexsurvspline}.}

\item{...}{Other arguments (currently unused).}
}
\value{
Log-likelihood (numeric) with additional attributes \code{df} (degrees of freedom, or number
of parameters that were estimated), and number of observations \code{nobs} (including observed
events and censored observations).
}
\description{
Log likelihood from a flexsurvreg model
}
