% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.flexsurvmix.R
\name{mean_flexsurvmix}
\alias{mean_flexsurvmix}
\title{Mean times to events from a flexsurvmix model}
\usage{
mean_flexsurvmix(x, newdata = NULL, B = NULL)
}
\arguments{
\item{x}{Fitted model object returned from \code{\link{flexsurvmix}}.}

\item{newdata}{Data frame or list of covariate values.}

\item{B}{Number of simulations to use to compute 95% confidence intervals,
based on the asymptotic multivariate normal distribution of the basic
parameter estimates.}
}
\value{
Mean times to next event conditionally on each alternative event,
  given the specified covariate values.
}
\description{
This returns the mean of each event-specific parametric time-to-event
distribution in the mixture model, which is the mean time to event
conditionally on that event being the one that happens.
}
