% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexsurvmix.R
\name{model.frame.flexsurvmix}
\alias{model.frame.flexsurvmix}
\title{Model frame from a flexsurvmix object}
\usage{
\method{model.frame}{flexsurvmix}(formula, ...)
}
\arguments{
\item{formula}{Fitted model object from \code{\link{flexsurvmix}}.}

\item{...}{Further arguments (currently unused).}
}
\value{
A list of data frames
}
\description{
Returns a list of data frames, with each component containing the
data that were used for the model fit for that mixture component.
}
