% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.flexsurvmix.R
\name{p_flexsurvmix}
\alias{p_flexsurvmix}
\title{Transition probabilities from a flexsurvmix model}
\usage{
p_flexsurvmix(x, newdata = NULL, t = 1, startname = "start", long = FALSE)
}
\arguments{
\item{x}{Fitted model object returned from \code{\link{flexsurvmix}}.}

\item{newdata}{Data frame or list of covariate values.}

\item{t}{Vector of times \code{t} to calculate the probabilities of
transition by.}

\item{startname}{Name of the state where individuals start. This considers
the model as a multi-state model where people start in this state, and may
transition to one of the competing events.}

\item{long}{If \code{TRUE} then the function returns a data frame in long
form.   The default is wide form, where probabilities for different states
are returned in different columns, and times and covariate values are in
different rows.  Long format has times, covariates and states in different
rows, which is ideal for plotting with \pkg{ggplot2}}
}
\value{
A data frame with transition probabilities by time, covariate value
  and destination state.
}
\description{
These quantities are variously known as transition probabilities, or state
occupancy probabilities, or values of the "cumulative incidence" function,
or values of the "subdistribution" function. They are the probabilities that
an individual has experienced an event of a particular kind by time
\code{t}.
}
\details{
Note that "cumulative incidence" is a misnomer, as "incidence" typically
means a hazard, and the quantities computed here are not cumulative hazards,
but probabilities.
}
