% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.flexsurvmix.R
\name{pdf_flexsurvmix}
\alias{pdf_flexsurvmix}
\title{Fitted densities for times to events in a flexsurvmix model}
\usage{
pdf_flexsurvmix(x, newdata = NULL, t = NULL)
}
\arguments{
\item{x}{Fitted model object returned from \code{\link{flexsurvmix}}.}

\item{newdata}{Data frame or list of covariate values.   If omitted for a
model with covariates, a default is used, defined by all combinations of
factors if the only covariates in the model are factors, or all covariate
values of zero if there are any non-factor covariates in the model.}

\item{t}{Vector of times at which to evaluate the probability density}
}
\value{
A data frame with each row giving the fitted density \code{dens} for
  a combination of covariate values, time and competing event.
}
\description{
This returns an estimate of the probability density for the time to each
competing event, at a vector of times supplied by the user.
}
