% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstate.R
\name{pfinal_fmsm}
\alias{pfinal_fmsm}
\title{Steady state probabilities in a flexible parametric Markov multi-state model}
\usage{
pfinal_fmsm(x, newdata = NULL, fromstate, maxt = 1e+05, B = 0, cores = NULL)
}
\arguments{
\item{x}{Object returned by \code{\link{fmsm}}, representing a multi-state
model formed from transition-specific time-to-event models fitted by
\code{\link{flexsurvreg}}.}

\item{newdata}{Data frame of covariate values, with one column per
covariate, and one row per alternative value.}

\item{fromstate}{State from which to calculate the transition probability
state.  This should refer to the name of a row of the transition matrix
\code{attr(x,trans)}.}

\item{maxt}{Large time to use for forecasting steady state probabilities.
The transition probability from zero to this time is used.  Note
\code{Inf} will not work. The default is \code{100000}.}

\item{B}{Number of simulations to use to calculate 95% confidence intervals
based on the asymptotic normal distribution of the basic parameter
estimates. If \code{B=0} then no intervals are calculated.}

\item{cores}{Number of processor cores to use.  If \code{NULL} (the default)
then a single core is used.}
}
\value{
A data frame with one row per covariate value and destination state,
  giving the state in column \code{state}, and probability in column
  \code{val}. Additional columns \code{lower} and \code{upper} for the
  confidence limits are returned if \code{B=0}.
}
\description{
This requires the model to be Markov, and is not valid for semi-Markov
models, as it works by wrapping \code{\link{pmatrix.fs}} to calculate the
transition probability over a very large time.  No check is performed for
the Markov property, so use with care.  For semi-Markov models, simulation
is required, e.g. through \code{\link{pmatrix.simfs}}.
}
\details{
Competing risks models, defined by just one starting state with multiple
destination states representing competing events, are Markov by definition.
For these models, this function returns the probability governing which
competing event happens next.
}
