% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survrtrunc.R
\name{plot.survrtrunc}
\alias{plot.survrtrunc}
\alias{lines.survrtrunc}
\title{Plot nonparametric estimates of survival from right-truncated data.}
\usage{
\method{plot}{survrtrunc}(x, ...)

\method{lines}{survrtrunc}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"survrtrunc"} as returned by \code{\link{survrtrunc}}.}

\item{...}{Other arguments to be passed to \code{\link[survival]{plot.survfit}} or \code{\link[survival]{lines.survfit}}.}
}
\description{
\code{plot.survrtrunc} creates a new plot, while \code{lines.survrtrunc} adds lines to an exising plot.
}
