% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmixmsm.R
\name{prob_pathway}
\alias{prob_pathway}
\title{Probability of each pathway taken through a mixture multi-state model}
\usage{
prob_pathway(x, newdata = NULL, final = FALSE, B = NULL)
}
\arguments{
\item{x}{Object returned by \code{\link{fmixmsm}}, representing a multi-state
model built from piecing together mixture models fitted by
\code{\link{flexsurvmix}}.}

\item{newdata}{Data frame or list of covariate values.}

\item{final}{If \code{TRUE} then the probabilities of pathways with the same
final state are added together, to produce the probability  of each
ultimate outcome or absorbing state from the multi-state model.}

\item{B}{Number of simulations to use to compute 95\% confidence intervals,
based on the asymptotic multivariate normal distribution of the basic
parameter estimates.  If \code{B=NULL} then intervals are not computed.}
}
\value{
Data frame of pathway probabilities by covariate value and pathway.
}
\description{
Probability of each pathway taken through a mixture multi-state model
}
