% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstate.R
\name{simfs_bytrans}
\alias{simfs_bytrans}
\title{Reformat simulated multi-state data with one row per simulated transition}
\usage{
simfs_bytrans(simfs)
}
\arguments{
\item{simfs}{Output from \code{\link{sim.fmsm}} representing simulated histories from a multi-state model.}
}
\value{
Data frame with four columns giving transition start state, transition end state, transition name and the time taken by the transition.
}
\description{
Reformat simulated multi-state data with one row per simulated transition
}
