% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmixmsm.R
\name{simt_flexsurvmix}
\alias{simt_flexsurvmix}
\title{Simulate times to competing events from a mixture multi-state model}
\usage{
simt_flexsurvmix(x, newdata = NULL, n)
}
\arguments{
\item{x}{Fitted model object returned from \code{\link{flexsurvmix}}.}

\item{newdata}{Data frame or list of covariate values.}

\item{n}{Number of simulations}
}
\value{
Data frame with \code{n*m} rows and a column for each competing
  event, where \code{m} is the number of alternative covariate values, that
  is the number of rows of \code{newdata}.   The simulated time represents
  the time to that event conditionally on that event being the one that
  occurs.  This function doesn't simulate which event occurs.
}
\description{
Simulate times to competing events from a mixture multi-state model
}
