% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standsurv.flexsurvreg.R
\name{standsurv.flexsurvreg}
\alias{standsurv.flexsurvreg}
\title{Marginal survival and hazards of fitted flexsurvreg models}
\usage{
standsurv.flexsurvreg(
  object,
  newdata = NULL,
  at = list(list()),
  atreference = 1,
  type = "survival",
  t = NULL,
  ci = FALSE,
  se = FALSE,
  B = 1000,
  cl = 0.95,
  contrast = NULL,
  seed = NULL
)
}
\arguments{
\item{object}{Output from \code{\link{flexsurvreg}} or
\code{\link{flexsurvspline}}, representing a fitted survival model object.}

\item{newdata}{Data frame containing covariate values to produce marginal
values for. If not specified then the fitted model data.frame is used.
There must be a column for every covariate in the model formula
for which the user wishes to standardize over.  These are in the same format
as the original data, with factors as a single variable, not 0/1 contrasts.
Any covariates that are to be fixed should be specified in \code{at}. 
There should be one row for every combination of covariates in which to 
standardize over.}

\item{at}{A list of scenarios in which specific covariates are fixed to 
certain values. Each element of \code{at} must itself be a list. For example,
for a covariate \code{group} with levels "Good", "Medium" and "Poor", the 
standardized survival plots for each group averaging over all other 
covariates is specified using 
\code{at=list(list(group="Good"), list(group="Medium"), list(group="Poor"))}.}

\item{atreference}{The reference scenario for making contrasts. Default is 1,
(i.e. the first element of \code{at}).}

\item{type}{\code{"survival"} for marginal survival probabilities.

\code{"hazard"} for the hazard of the marginal survival probability.

\code{"rmst"} for standardized restricted mean survival.}

\item{t}{Times to calculate marginal values at.}

\item{ci}{Should confidence intervals be calculated (using bootstrapping)? 
Defaults to FALSE}

\item{se}{Should standard errors be calculated (using bootstrapping)? 
Defaults to FALSE}

\item{B}{Number of simulations from the normal asymptotic distribution of the
estimates used to calculate confidence intervals or standard errors. Decrease
for greater speed at the expense of accuracy.}

\item{cl}{Width of symmetric confidence intervals, relative to 1.}

\item{contrast}{Contrasts between standardized measures defined by \code{at}
scenarios. Options are \code{"difference"} and \code{"ratio"}. There will be
n-1 new columns created where n is the number of \code{at} scenarios. Default
is NULL (i.e. no contrasts are calculated).}

\item{seed}{The random seed to use (for bootstrapping confidence intervals)}
}
\value{
A \code{tibble} containing one row for each 
time-point. The column naming convention is \code{at{i}} for the ith scenario
with corresponding confidence intervals (if specified) named \code{at{i}_lci}
and \code{at{i}_uci}. Contrasts are named \code{contrast{k}_{j}} for the 
comparison of the kth versus the jth \code{at} scenario.

In addition tidy long-format data.frames are returned in the attributes
\code{standsurv_at} and \code{standsurv_contrast}. These can be passed to 
\code{ggplot} for plotting purposes (see \code{\link{plot.standsurv}}).
}
\description{
Returns a tidy data.frame of marginal survival probabilities or the hazards 
of the marginal survival at user-defined time points and covariate patterns.
Standardization is performed over any undefined covariates in the model. 
The user provides the data to standardize over. Contrasts can be calculated 
resulting in estimates of the average treatment effect or the average 
treatment effect in the treated if a treated subset of the data are supplied.
}
\details{
The syntax of \code{standsurv.flexreg} follows closely that of Stata's 
\code{standsurv} command written by Paul Lambert and Michael Crowther. The 
function calculates standardized (marginal) measures including standardized
survival functions, standardized restricted mean survival times and the 
hazard of standardized survival. The standardized survival is defined as
\deqn{S_s(t|X=x) = E(S(t|X=x,Z)) = \frac{1}{N} \sum_{i=1}^N S(t|X=x,Z=z_i)}{S(t|X=x) = E[S(t|X=x,Z)] = 1/N * sum(S(t|X=x,Z=z_i))}
The hazard of the standardized survival is a weighted average of 
individual hazard functions at time t, weighted by the survival
function at this time:
\deqn{h_s(t|X=x) = \frac{\sum_{i=1}^N S(t|X=x,Z=z_i)h(t|X=x,Z=z_i)}{\sum_{i=1}^N S(t|X=x,Z=z_i)}}{h(t|X=x) = sum(S(t|X=x,Z=z_i) * h(t|X=x,Z=z_i)) / sum(S(t|X=x,Z=z_i))}
}
\examples{
## mean age is higher in those with smaller observed survival times 
newbc <- bc
newbc$age <- rnorm(dim(bc)[1], mean = 65-scale(newbc$recyrs, scale=FALSE),
 sd = 5)

## Fit a Weibull flexsurv model with group and age as covariates
weib_age <- flexsurvreg(Surv(recyrs, censrec) ~ group+age, data=newbc, 
                       dist="weibull")
                       
## Calculate standardized survival and the difference in standardized survival
## for the three levels of group across a grid of survival times                        
standsurv_weib_age <- standsurv.flexsurvreg(weib_age, 
                                           at = list(list(group="Good"), 
                                                     list(group="Medium"), 
                                                     list(group="Poor")), 
                                           t=seq(0,7, length=100),
                                           contrast = "difference", ci=FALSE)
standsurv_weib_age

## Calculate hazard of standardized survival and the marginal hazard ratio
## for the three levels of group across a grid of survival times
## 10 bootstraps for confidence intervals (this should be larger)          
haz_standsurv_weib_age <- standsurv.flexsurvreg(weib_age, 
                                           at = list(list(group="Good"), 
                                                     list(group="Medium"), 
                                                     list(group="Poor")), 
                                           t=seq(0,7, length=100),
                                           type="hazard",
                                           contrast = "ratio", B=10, ci=TRUE)
haz_standsurv_weib_age                                            
plot(haz_standsurv_weib_age, ci=TRUE)
## Hazard ratio plot shows a decreasing marginal HR 
## Whereas the conditional HR is constant (model is a PH model)
plot(haz_standsurv_weib_age, contrast=TRUE, ci=TRUE)
}
\references{
Paul Lambert, 2021. "STANDSURV: Stata module to compute 
standardized (marginal) survival and related functions," 
Statistical Software Components S458991, Boston College Department of 
Economics. https://ideas.repec.org/c/boc/bocode/s458991.html
}
\author{
Michael Sweeting <michael.sweeting@astrazeneca.com>
}
