% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standsurv.flexsurvreg.R
\name{tidy.standsurv}
\alias{tidy.standsurv}
\title{Tidy a standsurv object.}
\usage{
\method{tidy}{standsurv}(object)
}
\arguments{
\item{object}{A standsurv object.}
}
\value{
Returns additional tidy data.frames (tibbles)
stored as attributes named standpred_at and standpred_contrast.
}
\description{
This function is used internally by \code{standsurv.flexsurvreg} and tidy
data.frames are automatically returned by the function.
}
