% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexsurvreg.R
\name{vcov.flexsurvreg}
\alias{vcov.flexsurvreg}
\title{Variance-covariance matrix from  a flexsurvreg model}
\usage{
\method{vcov}{flexsurvreg}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class
\code{\link{flexsurvreg}}, e.g. as returned by
\code{flexsurvreg} or \code{flexsurvspline}.}

\item{...}{Other arguments (currently unused).}
}
\value{
Variance-covariance matrix of the estimated parameters, on
  the scale that they were estimated on (for positive parameters
  this is the log scale).
}
\description{
Variance-covariance matrix from  a flexsurvreg model
}
