% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeibullPH.R
\name{WeibullPH}
\alias{WeibullPH}
\alias{dweibullPH}
\alias{pweibullPH}
\alias{qweibullPH}
\alias{rweibullPH}
\alias{HweibullPH}
\alias{hweibullPH}
\title{Weibull distribution in proportional hazards parameterisation}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{p}{Vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is
taken to be the number required.}

\item{shape}{Vector of shape parameters.}

\item{scale}{Vector of scale parameters.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P(X
\le x)}{P(X <= x)}, otherwise, \eqn{P(X > x)}{P(X > x)}.}
}
\value{
\code{dweibullPH} gives the density, \code{pweibullPH} gives the
distribution function, \code{qweibullPH} gives the quantile function,
\code{rweibullPH} generates random deviates, \code{HweibullPH} retuns the
cumulative hazard and \code{hweibullPH} the hazard.
}
\description{
Density, distribution function, hazards, quantile function and random
generation for the Weibull distribution in its proportional hazards
parameterisation.
}
\details{
The Weibull distribution in proportional hazards parameterisation with
`shape' parameter a and `scale' parameter m has density given by

\deqn{f(x) = a m x^{a-1} exp(- m x^a) }

cumulative distribution function \eqn{F(x) = 1 - exp( -m x^a )}, survivor
function \eqn{S(x) = exp( -m x^a )}, cumulative hazard \eqn{m x^a} and
hazard \eqn{a m x^{a-1}}.

\code{\link{dweibull}} in base R has the alternative 'accelerated failure
time' (AFT) parameterisation with shape a and scale b.  The shape parameter
\eqn{a} is the same in both versions.  The scale parameters are related as
\eqn{b = m^{-1/a}}, equivalently m = b^-a.

In survival modelling, covariates are typically included through a linear
model on the log scale parameter.  Thus, in the proportional hazards model,
the coefficients in such a model on \eqn{m} are interpreted as log hazard
ratios.

In the AFT model, covariates on \eqn{b} are interpreted as time
acceleration factors.  For example, doubling the value of a covariate with
coefficient \eqn{beta=log(2)} would give half the expected survival time.
These coefficients are related to the log hazard ratios \eqn{\gamma} as
\eqn{\beta = -\gamma / a}.
}
\seealso{
\code{\link{dweibull}}
}
\author{
Christopher Jackson <chris.jackson@mrc-bsu.cam.ac.uk>
}
\keyword{distribution}
