% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexsurv-package.R
\name{means}
\alias{means}
\alias{mean_exp}
\alias{rmst_exp}
\alias{mean_weibull}
\alias{rmst_weibull}
\alias{mean_weibullPH}
\alias{rmst_weibullPH}
\alias{mean_llogis}
\alias{rmst_llogis}
\alias{mean_lnorm}
\alias{rmst_lnorm}
\alias{mean_gamma}
\alias{rmst_gamma}
\alias{mean_gompertz}
\alias{rmst_gompertz}
\alias{mean_gengamma}
\alias{rmst_gengamma}
\alias{mean_gengamma.orig}
\alias{rmst_gengamma.orig}
\alias{mean_genf}
\alias{rmst_genf}
\alias{mean_genf.orig}
\alias{rmst_genf.orig}
\title{Mean and restricted mean survival functions}
\arguments{
\item{t}{Vector of times to which rmst is evaluated}

\item{start}{Optional left-truncation time or times.  The returned
restricted mean survival will be conditioned on survival up to
this time.}

\item{rate}{Rate parameter (exponential and gamma)}

\item{shape}{Shape parameter (Weibull, gamma, log-logistic, generalized gamma [orig],
generalized F [orig])}

\item{scale}{Scale parameter (Weibull, log-logistic, generalized gamma [orig],
generalized F [orig])}

\item{meanlog}{Mean on the log scale (log normal)}

\item{sdlog}{Standard deviation on the log scale (log normal)}

\item{mu}{Mean on the log scale (generalized gamma, generalized F)}

\item{sigma}{Standard deviation on the log scale (generalized gamma, generalized F)}

\item{Q}{Vector of first shape parameters (generalized gamma, generalized F)}

\item{P}{Vector of second shape parameters (generalized F)}

\item{k}{vector of shape parameters (generalized gamma [orig]).}

\item{s1}{Vector of first F shape parameters (generalized F [orig])}

\item{s2}{vector of second F shape parameters (generalized F [orig])}
}
\value{
mean surival (functions beginning 'mean') or restricted mean survival
(functions beginning 'rmst_').
}
\description{
Mean and restricted mean survival functions for distributions which are built
into flexsurv, and whose distribution functions are in base R.
}
\details{
For the exponential, Weibull, log-logistic, lognormal, and gamma, mean survival is
provided analytically.  Restricted mean survival for the exponential distribution
is also provided analytically.  Mean and restricted means for other distributions
are calculated via numeric integration.
}
\seealso{
\code{\link{dexp}},\code{\link{dweibull}},\code{\link{dgamma}},\code{\link{dlnorm}},\code{\link{dgompertz}},\code{\link{dgengamma}},\code{\link{dgenf}}
}
\author{
Christopher Jackson <chris.jackson@mrc-bsu.cam.ac.uk>
}
\keyword{distribution}
