% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.flexsurvreg.R
\name{plot.flexsurvreg}
\alias{plot.flexsurvreg}
\title{Plots of fitted flexible survival models}
\usage{
\method{plot}{flexsurvreg}(x, newdata = NULL, X = NULL, type = "survival",
  fn = NULL, t = NULL, start = 0, est = TRUE, ci = NULL, B = 1000,
  cl = 0.95, col.obs = "black", lty.obs = 1, lwd.obs = 1, col = "red",
  lty = 1, lwd = 2, col.ci = NULL, lty.ci = 2, lwd.ci = 1,
  ylim = NULL, add = FALSE, ...)
}
\arguments{
\item{x}{Output from \code{\link{flexsurvreg}} or
\code{\link{flexsurvspline}}, representing a fitted survival model object.}

\item{newdata}{Data frame containing covariate values to produce fitted
values for.  See \code{\link{summary.flexsurvreg}}.

If there are only factor covariates in the model, then Kaplan-Meier (or
nonparametric hazard...)  curves are plotted for all distinct groups, and
by default, fitted curves are also plotted for these groups.  To plot
Kaplan-Meier and fitted curves for only a subset of groups, use
\code{plot(survfit())} followed by \code{lines.flexsurvreg()}.

If there are any continuous covariates, then a single population
Kaplan-Meier curve is drawn. By default, a single fitted curve is drawn
with the covariates set to their mean values in the data - for categorical
covariates, the means of the 0/1 indicator variables are taken.}

\item{X}{Alternative way to supply covariate values, as a model matrix.
See \code{\link{summary.flexsurvreg}}.  \code{newdata} is an easier way.}

\item{type}{\code{"survival"} for survival, to be plotted against
Kaplan-Meier estimates from \code{\link[survival]{plot.survfit}}.

\code{"cumhaz"} for cumulative hazard, plotted against transformed
Kaplan-Meier estimates from \code{\link[survival]{plot.survfit}}.

\code{"hazard"} for hazard, to be plotted against smooth nonparametric
estimates from \code{\link[muhaz]{muhaz}}.  The nonparametric estimates
tend to be unstable, and these plots are intended just to roughly indicate
the shape of the hazards through time.  The \code{min.time} and
\code{max.time} options to \code{\link[muhaz]{muhaz}} may sometimes need to
be passed as arguments to \code{\link{plot.flexsurvreg}} to avoid an error
here.

Ignored if \code{"fn"} is specified.}

\item{fn}{Custom function of the parameters to summarise against time.  The
first two arguments of the function must be \code{t} representing time, and
\code{start} representing left-truncation points, and any remaining
arguments must be parameters of the distribution.  It should return a
vector of the same length as \code{t}.}

\item{t}{Vector of times to plot fitted values for, see
\code{\link{summary.flexsurvreg}}.}

\item{start}{Left-truncation points, see \code{\link{summary.flexsurvreg}}.}

\item{est}{Plot fitted curves (\code{TRUE} or \code{FALSE}.)}

\item{ci}{Plot confidence intervals for fitted curves. By default, this is
\code{TRUE} if one observed/fitted curve is plotted, and \code{FALSE} if
multiple curves are plotted.}

\item{B}{Number of simulations controlling accuracy of confidence
intervals, as used in \code{\link[=summary.flexsurvreg]{summary}}.
Decrease for greater speed at the expense of accuracy, or set \code{B=0} to
turn off calculation of CIs.}

\item{cl}{Width of confidence intervals, by default 0.95 for 95\%
intervals.}

\item{col.obs}{Colour of the nonparametric curve.}

\item{lty.obs}{Line type of the nonparametric curve.}

\item{lwd.obs}{Line width of the nonparametric curve.}

\item{col}{Colour of the fitted parametric curve(s).}

\item{lty}{Line type of the fitted parametric curve(s).}

\item{lwd}{Line width of the fitted parametric curve(s).}

\item{col.ci}{Colour of the fitted confidence limits, defaulting to the
same as for the fitted curve.}

\item{lty.ci}{Line type of the fitted confidence limits.}

\item{lwd.ci}{Line width of the fitted confidence limits.}

\item{ylim}{y-axis limits: vector of two elements.}

\item{add}{If \code{TRUE}, add lines to an existing plot, otherwise new
axes are drawn.}

\item{...}{Other options to be passed to \code{\link{plot.survfit}} or
\code{\link[muhaz]{muhaz}}, for example, to control the smoothness of the
nonparametric hazard estimates.  The \code{min.time} and \code{max.time}
options to \code{\link[muhaz]{muhaz}} may sometimes need to be changed from
the defaults.}
}
\description{
Plot fitted survival, cumulative hazard or hazard from a parametric model
against nonparametric estimates to diagnose goodness-of-fit.  Alternatively
plot a user-defined function of the model parameters against time.
}
\note{
Some standard plot arguments such as \code{"xlim","xlab"} may not
work.  This function was designed as a quick check of model fit.  Users
wanting publication-quality graphs are advised to set up an empty plot with
the desired axes first (e.g. with \code{plot(...,type="n",...)}), then use
suitable \code{\link{lines}} functions to add lines.
}
\seealso{
\code{\link{flexsurvreg}}
}
\author{
C. H. Jackson \email{chris.jackson@mrc-bsu.cam.ac.uk}
}
\keyword{hplot}
\keyword{models}
